[FLINK-14260] Replace TestingJobManagerRunnerFactory with TestingJobManagerRunnerFactoryNG
This commit removes the old TestingJobManagerRunnerFactory and renames
the TestingJobManagerRunnerFactoryNG into TestingJobManagerRunnerFactory.

This closes #9789.
