+++ flink-runtime/src/test/java/org/apache/flink/runtime/dispatcher/TestingJobManagerRunnerFactory.java
@@ -19,76 +19,68 @@
 package org.apache.flink.runtime.dispatcher;
 
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.runtime.executiongraph.ArchivedExecutionGraph;
 import org.apache.flink.runtime.heartbeat.HeartbeatServices;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.jobgraph.JobGraph;
-import org.apache.flink.runtime.jobmaster.JobManagerRunner;
-import org.apache.flink.runtime.jobmaster.JobManagerRunnerImpl;
 import org.apache.flink.runtime.jobmaster.JobManagerSharedServices;
+import org.apache.flink.runtime.jobmaster.TestingJobManagerRunner;
 import org.apache.flink.runtime.jobmaster.factories.JobManagerJobMetricGroupFactory;
 import org.apache.flink.runtime.rpc.FatalErrorHandler;
 import org.apache.flink.runtime.rpc.RpcService;
 
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Supplier;
+import javax.annotation.Nonnull;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
+import java.util.concurrent.ArrayBlockingQueue;
+import java.util.concurrent.BlockingQueue;
 
 /**
- * {@link JobManagerRunnerFactory} implementation for
- * testing purposes.
+ * Testing implementation of {@link JobManagerRunnerFactory} which returns a {@link TestingJobManagerRunner}.
  */
 class TestingJobManagerRunnerFactory implements JobManagerRunnerFactory {
 
-	private final CompletableFuture<JobGraph> jobGraphFuture;
-	private final CompletableFuture<ArchivedExecutionGraph> resultFuture;
-	private final CompletableFuture<Void> terminationFuture;
-	private final AtomicReference<Supplier<Exception>> failJobMasterCreationWith;
+	private final BlockingQueue<TestingJobManagerRunner> createdJobManagerRunner = new ArrayBlockingQueue<>(16);
 
-	TestingJobManagerRunnerFactory(
-			CompletableFuture<JobGraph> jobGraphFuture,
-			CompletableFuture<ArchivedExecutionGraph> resultFuture,
-			CompletableFuture<Void> terminationFuture) {
-		this(jobGraphFuture, resultFuture, terminationFuture, new AtomicReference<>());
+	private int numBlockingJobManagerRunners;
+
+	public TestingJobManagerRunnerFactory() {
+		this(0);
 	}
 
-	TestingJobManagerRunnerFactory(
-			CompletableFuture<JobGraph> jobGraphFuture,
-			CompletableFuture<ArchivedExecutionGraph> resultFuture,
-			CompletableFuture<Void> terminationFuture,
-			AtomicReference<Supplier<Exception>> failJobMasterCreationWith) {
-		this.jobGraphFuture = jobGraphFuture;
-		this.resultFuture = resultFuture;
-		this.terminationFuture = terminationFuture;
-		this.failJobMasterCreationWith = failJobMasterCreationWith;
+	public TestingJobManagerRunnerFactory(int numBlockingJobManagerRunners) {
+		this.numBlockingJobManagerRunners = numBlockingJobManagerRunners;
 	}
 
 	@Override
-	public JobManagerRunner createJobManagerRunner(
+	public TestingJobManagerRunner createJobManagerRunner(
 			JobGraph jobGraph,
 			Configuration configuration,
 			RpcService rpcService,
 			HighAvailabilityServices highAvailabilityServices,
 			HeartbeatServices heartbeatServices,
-			JobManagerSharedServices jobManagerSharedServices,
+			JobManagerSharedServices jobManagerServices,
 			JobManagerJobMetricGroupFactory jobManagerJobMetricGroupFactory,
 			FatalErrorHandler fatalErrorHandler) throws Exception {
-		final Supplier<Exception> exceptionSupplier = failJobMasterCreationWith.get();
+		final TestingJobManagerRunner testingJobManagerRunner = createTestingJobManagerRunner(jobGraph);
+		createdJobManagerRunner.offer(testingJobManagerRunner);
 
-		if (exceptionSupplier != null) {
-			throw exceptionSupplier.get();
+		return testingJobManagerRunner;
+	}
+
+	@Nonnull
+	private TestingJobManagerRunner createTestingJobManagerRunner(JobGraph jobGraph) {
+		final boolean blockingTermination;
+
+		if (numBlockingJobManagerRunners > 0) {
+			numBlockingJobManagerRunners--;
+			blockingTermination = true;
 		} else {
-			jobGraphFuture.complete(jobGraph);
-
-			final JobManagerRunnerImpl mock = mock(JobManagerRunnerImpl.class);
-			when(mock.getResultFuture()).thenReturn(resultFuture);
-			when(mock.closeAsync()).thenReturn(terminationFuture);
-			when(mock.getJobID()).thenReturn(jobGraph.getJobID());
-
-			return mock;
+			blockingTermination = false;
 		}
+
+		return new TestingJobManagerRunner(jobGraph.getJobID(), blockingTermination);
+	}
+
+	public TestingJobManagerRunner takeCreatedJobManagerRunner() throws InterruptedException {
+		return createdJobManagerRunner.take();
 	}
 }
